# api/v1/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  req:
    description: Our base64 document
    properties:
      base64:
        type: string
        x-go-name: Base
      token:
        type: string
        x-go-name: Token
    type: object
    x-go-package: github.com/JetBrainer/BackOCRService/internal/app/apiserver
  request:
    properties:
      email:
        type: string
        x-go-name: Email
      organization:
        type: string
        x-go-name: Organization
      password:
        type: string
        x-go-name: Password
    type: object
    x-go-package: github.com/JetBrainer/BackOCRService/internal/app/apiserver
  response:
    properties:
      id:
        format: int64
        type: integer
        x-go-name: ID
      token:
        type: string
        x-go-name: Token
    type: object
    x-go-package: github.com/JetBrainer/BackOCRService/internal/app/apiserver
info:
  description: Documentation for Enterprise Intelligent Character Recognition API
  title: of Document API
  version: 1.0.0
paths:
  /image:
    post:
      description: |-
        Document return

        Client sends Full Scanned Document and get's every need field
      operationId: docRequest
      parameters:
      - description: Need data
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/req'
      responses:
        "200":
          $ref: '#/responses/docResponse'
      summary: Returns particular document field
      tags:
      - Document
  /register:
    post:
      description: |-
        account

        User creates account to get Token
      operationId: formCreateReq
      parameters:
      - description: form request
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/request'
      responses:
        "200":
          $ref: '#/responses/tokenResponse'
      summary: Returns Id and Token for OCR
      tags:
      - Account
produces:
- application/json
responses:
  docResponse:
    description: Your Invoice structure
    schema:
      items:
        type: object
      type: array
  tokenResponse:
    description: ID and Token
    schema:
      items:
        $ref: '#/definitions/response'
      type: array
schemes:
- http
swagger: "2.0"